+++ syncrclone/utils.py
@@ -144,8 +144,8 @@ def search_upwards(pwd):
     return search_upwards(newpwd)
         
         
-        
-        
+            
+