+++ tests/run_with_cov.sh
@@ -1,0 +1,1 @@
+pytest --cov=syncrclone --cov-report html test.py